requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn==0.29.0
# File parsing
python-docx==1.1.0
PyMuPDF==1.23.24         
# NLP & ML
spacy==3.7.2
scikit-learn==1.4.2
pandas==2.2.2
# Resume parsing and data handling
pdfminer.six==20231228   # Alternate option to PyMuPDF
numpy==1.26.4

python-multipart==0.0.9  # Needed for file uploads
pydantic==2.7.0          # Data validation

google-generativeai
python-dotenv